fix(e2e): share worker-scoped fixtures across test files to prevent browser context conflicts
