+++ apps/extension/tests/fixtures/shortcuts-fixture.ts
@@ -1,16 +1,7 @@
 /* eslint-disable react-hooks/rules-of-hooks */
-import {
-  type BrowserContext,
-  type Page,
-  type Worker,
-  test as base,
-} from '@playwright/test';
-import {
-  authenticateAndNavigate,
-  createSharedBackgroundSW,
-  createSharedContext,
-  getExtensionId,
-} from './base-fixture';
+import { type BrowserContext, type Page, type Worker } from '@playwright/test';
+import { authenticateAndNavigate } from './base-fixture';
+import { test as base } from './extension-base-fixture';
 
 export const test = base.extend<
   {
@@ -18,47 +9,25 @@ export const test = base.extend<
     context: BrowserContext;
   },
   {
-    sharedContext: BrowserContext;
-    sharedBackgroundSW: Worker;
-    sharedExtensionId: string;
     sharedPage: Page;
+    _sharedBackgroundSW: Worker;
   }
 >({
-  sharedContext: [
-    // eslint-disable-next-line no-empty-pattern
-    async ({}, use) => {
-      const { browserContext, userDataDir } = await createSharedContext();
-      await use(browserContext);
-      await browserContext.close();
-      const fsPromises = await import('node:fs/promises');
-      await fsPromises.rm(userDataDir, { recursive: true, force: true });
-    },
-    { scope: 'worker' },
-  ],
-
-  sharedBackgroundSW: [
-    async ({ sharedContext }, use) => {
-      const background = await createSharedBackgroundSW(sharedContext);
-      await use(background);
-    },
-    { scope: 'worker' },
-  ],
-
-  sharedExtensionId: [
+  // Include sharedBackgroundSW in worker scope to pass to authenticateAndNavigate
+  _sharedBackgroundSW: [
     async ({ sharedBackgroundSW }, use) => {
-      const id = await getExtensionId(sharedBackgroundSW);
-      await use(id);
+      await use(sharedBackgroundSW);
     },
     { scope: 'worker' },
   ],
 
   sharedPage: [
-    async ({ sharedContext, sharedExtensionId, sharedBackgroundSW }, use) => {
+    async ({ sharedContext, sharedExtensionId, _sharedBackgroundSW }, use) => {
       const page = await authenticateAndNavigate(
         sharedContext,
         sharedExtensionId,
         'shortcuts',
-        sharedBackgroundSW
+        _sharedBackgroundSW
       );
       await use(page);
     },
@@ -74,4 +43,4 @@ export const test = base.extend<
   },
 });
 
-export const { expect } = test;
+export { expect } from './extension-base-fixture';
